+++ src/io/segment/android/info/Device.java
@@ -5,8 +5,6 @@ import io.segment.android.models.EasyJSONObject;
 import org.json.JSONObject;
 
 import android.content.Context;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager.NameNotFoundException;
 
 public class Device implements Info<JSONObject> {
 
@@ -20,21 +18,12 @@ public class Device implements Info<JSONObject> {
 
 		EasyJSONObject device = new EasyJSONObject();
 		
-		PackageInfo packageInfo;
-		try {
-			packageInfo = context.getPackageManager().getPackageInfo(
-					context.getPackageName(), 0);
-			device.put("versionCode", packageInfo.versionCode);
-			device.put("versionName", packageInfo.versionName);
-			
-		} catch (NameNotFoundException e) {}
-		
 		device.put("sdk", android.os.Build.VERSION.SDK_INT);
 		device.put("release",  android.os.Build.VERSION.RELEASE);
 		device.put("brand", android.os.Build.BRAND);
 		device.put("manufacturer",  android.os.Build.MANUFACTURER);
 		
-		return device;
+		return device;	
 	}
 
 }