+++ history.md
@@ -1,3 +1,8 @@
+0.1.1 / 2013-04-23
+=================
+* Moving context.build out of context.device
+* Fixing global context not propagating bug
+
 0.1.0 / 2013-04-12
 =================
 * API stabilized